create product controller
