+++ src/routes/productsRouter.ts
@@ -1,53 +1,17 @@
-import express, { NextFunction, Request, Response } from "express";
-import { productService } from "../services";
-import { NotFoundError } from "../error-handler/custom-errors";
-import { ErrorCodes } from "../error-handler/errorCodes";
-import { param } from "express-validator";
-
-interface ProductsRequestParams {
-  id: number;
-}
+import express from "express";
+import productsController from "../controllers/productsController";
+import { GetItemByIdInParamsValidationSchema } from "../middleware/validationSchemas";
+import { validationMiddleware } from "../middleware/validationMiddleware";
 
 const productsRouter = express.Router();
 
-productsRouter.get(
-  "/",
-  async function (req: Request, res: Response, next: NextFunction) {
-    try {
-      const products = await productService.getProducts();
-
-      res.json(products);
-    } catch (e) {
-      next(e);
-    }
-  }
-);
+productsRouter.get("/", productsController.getProducts);
 
 productsRouter.get(
   "/:id",
-  param("id").toInt(),
-  async function (
-    req: Request<ProductsRequestParams>,
-    res: Response,
-    next: NextFunction
-  ) {
-    const productId = req.params.id;
-
-    try {
-      const product = await productService.getProductById(productId);
-
-      if (!product) {
-        throw new NotFoundError({
-          message: `Product with id ${productId} was not found`,
-          code: ErrorCodes.PRODUCT_NOT_FOUND,
-        });
-      }
-
-      res.json(product);
-    } catch (e) {
-      next(e);
-    }
-  }
+  GetItemByIdInParamsValidationSchema,
+  validationMiddleware,
+  productsController.getProductById
 );
 
 export default productsRouter;
